requirements.txt
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.11.0
statsmodels>=0.14.0
scikit-learn>=1.4.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
matplotlib>=3.8.0
python-dotenv>=1.0.0
pytest>=8.0.0
boto3>=1.34.0
streamlit>=1.32.0
